requirements.txt
aiohttp==1.1.6
async-timeout==1.1.0
blessings==1.6
boto==2.43.0
bpython==0.16
chardet==2.3.0
coverage==4.2
curtsies==0.2.11
defusedxml==0.4.1
dj-database-url==0.4.1
dj-email-url==0.0.10
Django==1.10.3
django-allauth==0.29.0
django-debug-toolbar==1.6
django-filter==1.0.1
django-model-utils==2.6
django-post-office==2.0.8
django-redis-cache==1.7.1
django-rest-auth==0.8.2
django-storages==1.5.1
django-widget-tweaks==1.4.1
djangorestframework==3.5.3
factory-boy==2.7.0
fake-factory==0.7.2
greenlet==0.4.10
gunicorn==19.6.0
invoke==0.13.0
jsonfield==1.0.3
multidict==2.1.3
oauthlib==2.0.1
Pillow==3.4.2
psycopg2==2.6.2
Pygments==2.1.3
pyinotify==0.9.6
python-coveralls==2.9.0
python-dateutil==2.6.0
python3-openid==3.0.10
pytz==2016.7
PyYAML==3.12
redis==2.10.5
requests==2.12.1
requests-oauthlib==0.7.0
selenium==3.0.2
six==1.10.0
sqlparse==0.2.2
wcwidth==0.1.7
whitenoise==3.2.2
yarl==0.7.1
